refactor: split CharacterService into specialized modules
- Split monolithic CharacterService into 4 focused modules
- CharacterServiceCRUD: CRUD operations (270 lines)
- CharacterServiceSearch: Search and filtering (242 lines)
- CharacterServiceStats: Statistics calculations (239 lines)
- CharacterServiceTemplates: Templates and bulk operations (270 lines)
- Main CharacterService now acts as facade (382 lines)
- All modules under 500 lines to meet Codacy requirements
- Maintains same API through delegation pattern

Resolves Codacy complexity warnings while preserving functionality.
