+++ src/lib/services/CharacterService.ts
@@ -1,12 +1,9 @@
 /**
  * Character Service Layer for D&D Encounter Tracker
  *
- * Provides business logic for character management operations.
- * Abstracts database operations from API routes and provides
- * consistent error handling and validation.
- *
- * This class acts as a coordination layer, delegating operations
- * to specialized modules for better organization and maintainability.
+ * Main coordination layer that delegates operations to specialized modules.
+ * This class acts as a facade to maintain a clean API while keeping
+ * individual modules focused and under 500 lines each.
  */
 
 import { Types } from 'mongoose';
@@ -26,63 +23,26 @@ import {
   createErrorResult,
   CharacterServiceErrors,
 } from './CharacterServiceErrors';
+import { characterCreationSchema } from '../validations/character';
+
+// Import specialized service modules
+import { CharacterServiceCRUD } from './CharacterServiceCRUD';
 import {
-  characterCreationSchema,
-  characterUpdateSchema,
-} from '../validations/character';
-
-// Types for service operations
-export interface PaginatedCharacters {
-  items: ICharacter[];
-  pagination: {
-    page: number;
-    limit: number;
-    total: number;
-    totalPages: number;
-  };
-}
-
-export interface CharacterStats {
-  abilityModifiers: Record<string, number>;
-  savingThrows: Record<string, number>;
-  skills: Record<string, number>;
-  totalLevel: number;
-  classLevels: Record<string, number>;
-  proficiencyBonus: number;
-  initiativeModifier: number;
-  armorClass: number;
-  effectiveHitPoints: number;
-  status: 'alive' | 'unconscious' | 'dead';
-  isAlive: boolean;
-  isUnconscious: boolean;
-}
-
-export interface SpellcastingStats {
-  casterLevel: number;
-  spellSlots: Record<number, number>;
-  spellAttackBonus: number;
-  spellSaveDC: number;
-}
-
-export interface CarryingCapacity {
-  maximum: number;
-  current: number;
-  encumbranceLevel: 'none' | 'light' | 'heavy' | 'overloaded';
-}
-
-export interface EquipmentWeight {
-  total: number;
-  equipped: number;
-  carried: number;
-}
-
-export interface ExperienceInfo {
-  currentXP: number;
-  currentLevel: number;
-  nextLevelXP: number;
-  xpToNextLevel: number;
-  isMaxLevel: boolean;
-}
+  CharacterServiceSearch,
+  type PaginatedCharacters
+} from './CharacterServiceSearch';
+import {
+  CharacterServiceStats,
+  type CharacterStats,
+  type SpellcastingStats,
+  type CarryingCapacity,
+  type EquipmentWeight,
+  type ExperienceInfo,
+} from './CharacterServiceStats';
+import {
+  CharacterServiceTemplates,
+  type BulkOperationResult,
+} from './CharacterServiceTemplates';
 
 export interface CharacterPermissions {
   canView: boolean;
@@ -91,662 +51,186 @@ export interface CharacterPermissions {
   canShare: boolean;
 }
 
-export interface BulkOperationResult<T> {
-  successful: T[];
-  failed: Array<{ data: any; error: string }>;
-}
-
 /**
- * Character Service Layer
+ * Character Service Layer - Main Facade
  */
 export class CharacterService {
   // ================================
-  // CRUD Operations
+  // CRUD Operations - Delegate to CharacterServiceCRUD
   // ================================
 
-  /**
-   * Create a new character
-   */
   static async createCharacter(
     ownerId: string,
     characterData: CharacterCreation
   ): Promise<ServiceResult<ICharacter>> {
-    try {
-      // Validate owner ID
-      if (!Types.ObjectId.isValid(ownerId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(ownerId));
-      }
-
-      // Validate character data
-      const validationResult = characterCreationSchema.safeParse(characterData);
-      if (!validationResult.success) {
-        return createErrorResult(
-          CharacterServiceErrors.invalidCharacterData(validationResult.error.errors)
-        );
-      }
-
-      const validatedData = validationResult.data;
-
-      // Check character limit for subscription tier (mock implementation)
-      const characterCount = await Character.countDocuments({ ownerId: new Types.ObjectId(ownerId) });
-      const maxCharacters = await this.getCharacterLimitForUser(ownerId);
-
-      if (characterCount >= maxCharacters) {
-        return createErrorResult(
-          CharacterServiceErrors.characterLimitExceeded(characterCount, maxCharacters)
-        );
-      }
-
-      // Create character
-      const character = new Character({
-        ...validatedData,
-        ownerId: new Types.ObjectId(ownerId),
-      });
-
-      const savedCharacter = await character.save();
-      return createSuccessResult(savedCharacter);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('create character', error)
-      );
-    }
+    return CharacterServiceCRUD.createCharacter(ownerId, characterData);
   }
 
-  /**
-   * Get character by ID
-   */
   static async getCharacterById(
     characterId: string,
     userId: string
   ): Promise<ServiceResult<ICharacter>> {
-    try {
-      // Validate IDs
-      if (!Types.ObjectId.isValid(characterId)) {
-        return createErrorResult(CharacterServiceErrors.invalidCharacterId(characterId));
-      }
-      if (!Types.ObjectId.isValid(userId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(userId));
-      }
-
-      // Find character
-      const character = await Character.findById(characterId);
-      if (!character) {
-        return createErrorResult(CharacterServiceErrors.characterNotFound(characterId));
-      }
-
-      // Check access permissions
-      const hasAccess = await this.checkCharacterAccess(characterId, userId);
-      if (!hasAccess.success) {
-        return hasAccess;
-      }
-
-      return createSuccessResult(character);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get character', error)
-      );
-    }
+    return CharacterServiceCRUD.getCharacterById(characterId, userId);
   }
 
-  /**
-   * Update character
-   */
   static async updateCharacter(
     characterId: string,
     userId: string,
     updateData: CharacterUpdate
   ): Promise<ServiceResult<ICharacter>> {
-    try {
-      // Validate IDs
-      if (!Types.ObjectId.isValid(characterId)) {
-        return createErrorResult(CharacterServiceErrors.invalidCharacterId(characterId));
-      }
-
-      // Validate update data
-      const validationResult = characterUpdateSchema.safeParse(updateData);
-      if (!validationResult.success) {
-        return createErrorResult(
-          CharacterServiceErrors.invalidCharacterData(validationResult.error.errors)
-        );
-      }
-
-      // Check ownership
-      const ownershipCheck = await this.checkCharacterOwnership(characterId, userId);
-      if (!ownershipCheck.success) {
-        return ownershipCheck;
-      }
-
-      // Update character
-      const character = await Character.findByIdAndUpdate(
-        characterId,
-        { ...validationResult.data, updatedAt: new Date() },
-        { new: true, runValidators: true }
-      );
-
-      if (!character) {
-        return createErrorResult(CharacterServiceErrors.characterNotFound(characterId));
-      }
-
-      return createSuccessResult(character);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('update character', error)
-      );
-    }
+    return CharacterServiceCRUD.updateCharacter(characterId, userId, updateData);
   }
 
-  /**
-   * Delete character
-   */
   static async deleteCharacter(
     characterId: string,
     userId: string
   ): Promise<ServiceResult<void>> {
-    try {
-      // Validate IDs
-      if (!Types.ObjectId.isValid(characterId)) {
-        return createErrorResult(CharacterServiceErrors.invalidCharacterId(characterId));
-      }
-
-      // Check ownership
-      const ownershipCheck = await this.checkCharacterOwnership(characterId, userId);
-      if (!ownershipCheck.success) {
-        return ownershipCheck;
-      }
-
-      // Check if character is in use
-      const inUseCheck = await this.checkCharacterInUse(characterId);
-      if (!inUseCheck.success) {
-        return inUseCheck;
-      }
-
-      // Delete character
-      const character = await Character.findByIdAndDelete(characterId);
-      if (!character) {
-        return createErrorResult(CharacterServiceErrors.characterNotFound(characterId));
-      }
-
-      return createSuccessResult(void 0);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('delete character', error)
-      );
-    }
+    return CharacterServiceCRUD.deleteCharacter(characterId, userId);
   }
 
   // ================================
-  // Search and Filtering Operations
+  // Search and Filtering - Delegate to CharacterServiceSearch
   // ================================
 
-  /**
-   * Get characters by owner with pagination
-   */
   static async getCharactersByOwner(
     ownerId: string,
     page: number = 1,
     limit: number = 20
   ): Promise<ServiceResult<PaginatedCharacters>> {
-    try {
-      // Validate owner ID
-      if (!Types.ObjectId.isValid(ownerId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(ownerId));
-      }
-
-      const skip = (page - 1) * limit;
-      const ownerObjectId = new Types.ObjectId(ownerId);
-
-      const [characters, total] = await Promise.all([
-        Character.find({ ownerId: ownerObjectId })
-          .sort({ name: 1 })
-          .skip(skip)
-          .limit(limit),
-        Character.countDocuments({ ownerId: ownerObjectId }),
-      ]);
-
-      return createSuccessResult({
-        items: characters,
-        pagination: {
-          page,
-          limit,
-          total,
-          totalPages: Math.ceil(total / limit),
-        },
-      });
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get characters by owner', error)
-      );
-    }
+    return CharacterServiceSearch.getCharactersByOwner(ownerId, page, limit);
   }
 
-  /**
-   * Search characters by name
-   */
   static async searchCharacters(
     searchTerm: string,
     userId: string
   ): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      if (!searchTerm || searchTerm.trim().length === 0) {
-        return createErrorResult(
-          CharacterServiceErrors.invalidSearchCriteria({ searchTerm })
-        );
-      }
-
-      if (!Types.ObjectId.isValid(userId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(userId));
-      }
-
-      const userObjectId = new Types.ObjectId(userId);
-
-      // Search in owned characters and public characters
-      const characters = await Character.find({
-        $and: [
-          {
-            $or: [
-              { ownerId: userObjectId },
-              { isPublic: true },
-            ],
-          },
-          {
-            $text: { $search: searchTerm },
-          },
-        ],
-      }).sort({ score: { $meta: 'textScore' } });
-
-      return createSuccessResult(characters);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('search characters', error)
-      );
-    }
+    return CharacterServiceSearch.searchCharacters(searchTerm, userId);
   }
 
-  /**
-   * Get characters by class
-   */
   static async getCharactersByClass(
     className: CharacterClass,
     userId: string
   ): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      if (!Types.ObjectId.isValid(userId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(userId));
-      }
-
-      const userObjectId = new Types.ObjectId(userId);
-
-      const characters = await Character.find({
-        $and: [
-          {
-            $or: [
-              { ownerId: userObjectId },
-              { isPublic: true },
-            ],
-          },
-          {
-            'classes.class': className,
-          },
-        ],
-      }).sort({ name: 1 });
-
-      return createSuccessResult(characters);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get characters by class', error)
-      );
-    }
+    return CharacterServiceSearch.getCharactersByClass(className, userId);
   }
 
-  /**
-   * Get characters by race
-   */
   static async getCharactersByRace(
     race: CharacterRace,
     userId: string
   ): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      if (!Types.ObjectId.isValid(userId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(userId));
-      }
-
-      const userObjectId = new Types.ObjectId(userId);
-
-      const characters = await Character.find({
-        $and: [
-          {
-            $or: [
-              { ownerId: userObjectId },
-              { isPublic: true },
-            ],
-          },
-          { race },
-        ],
-      }).sort({ name: 1 });
-
-      return createSuccessResult(characters);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get characters by race', error)
-      );
-    }
+    return CharacterServiceSearch.getCharactersByRace(race, userId);
   }
 
-  /**
-   * Get characters by type
-   */
   static async getCharactersByType(
     type: CharacterType,
     userId: string
   ): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      if (!Types.ObjectId.isValid(userId)) {
-        return createErrorResult(CharacterServiceErrors.invalidOwnerId(userId));
-      }
-
-      const userObjectId = new Types.ObjectId(userId);
-
-      const characters = await Character.find({
-        $and: [
-          {
-            $or: [
-              { ownerId: userObjectId },
-              { isPublic: true },
-            ],
-          },
-          { type },
-        ],
-      }).sort({ name: 1 });
-
-      return createSuccessResult(characters);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get characters by type', error)
-      );
-    }
+    return CharacterServiceSearch.getCharactersByType(type, userId);
   }
 
-  /**
-   * Get public characters
-   */
   static async getPublicCharacters(): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      const characters = await Character.find({ isPublic: true })
-        .sort({ name: 1 });
-
-      return createSuccessResult(characters);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get public characters', error)
-      );
-    }
+    return CharacterServiceSearch.getPublicCharacters();
   }
 
   // ================================
-  // Character Statistics and Calculations
+  // Statistics and Calculations - Delegate to CharacterServiceStats
   // ================================
 
-  /**
-   * Calculate character statistics
-   */
   static async calculateCharacterStats(
     characterId: string,
     userId: string
   ): Promise<ServiceResult<CharacterStats>> {
-    try {
-      const characterResult = await this.getCharacterById(characterId, userId);
-      if (!characterResult.success) {
-        return characterResult;
-      }
-
-      const character = characterResult.data;
-
-      // Calculate ability modifiers
-      const abilityModifiers = {
-        strength: character.getAbilityModifier('strength'),
-        dexterity: character.getAbilityModifier('dexterity'),
-        constitution: character.getAbilityModifier('constitution'),
-        intelligence: character.getAbilityModifier('intelligence'),
-        wisdom: character.getAbilityModifier('wisdom'),
-        charisma: character.getAbilityModifier('charisma'),
-      };
-
-      // Calculate saving throws
-      const savingThrows = {
-        strength: abilityModifiers.strength + (character.savingThrows.strength ? character.proficiencyBonus : 0),
-        dexterity: abilityModifiers.dexterity + (character.savingThrows.dexterity ? character.proficiencyBonus : 0),
-        constitution: abilityModifiers.constitution + (character.savingThrows.constitution ? character.proficiencyBonus : 0),
-        intelligence: abilityModifiers.intelligence + (character.savingThrows.intelligence ? character.proficiencyBonus : 0),
-        wisdom: abilityModifiers.wisdom + (character.savingThrows.wisdom ? character.proficiencyBonus : 0),
-        charisma: abilityModifiers.charisma + (character.savingThrows.charisma ? character.proficiencyBonus : 0),
-      };
-
-      // Calculate skills
-      const skills: Record<string, number> = {};
-      for (const [skill, isProficient] of character.skills.entries()) {
-        const abilityMod = this.getSkillAbilityModifier(skill, abilityModifiers);
-        skills[skill] = abilityMod + (isProficient ? character.proficiencyBonus : 0);
-      }
-
-      // Calculate class levels
-      const classLevels: Record<string, number> = {};
-      character.classes.forEach(cls => {
-        classLevels[cls.class] = cls.level;
-      });
-
-      const stats: CharacterStats = {
-        abilityModifiers,
-        savingThrows,
-        skills,
-        totalLevel: character.level,
-        classLevels,
-        proficiencyBonus: character.proficiencyBonus,
-        initiativeModifier: character.getInitiativeModifier(),
-        armorClass: character.armorClass,
-        effectiveHitPoints: character.getEffectiveHP(),
-        status: character.isAlive() ? 'alive' : (character.isUnconscious() ? 'unconscious' : 'dead'),
-        isAlive: character.isAlive(),
-        isUnconscious: character.isUnconscious(),
-      };
-
-      return createSuccessResult(stats);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('calculate character stats', error)
-      );
-    }
+    return CharacterServiceStats.calculateCharacterStats(characterId, userId);
   }
 
-  /**
-   * Get character summary
-   */
   static async getCharacterSummary(
     characterId: string,
     userId: string
   ): Promise<ServiceResult<CharacterSummary>> {
-    try {
-      const characterResult = await this.getCharacterById(characterId, userId);
-      if (!characterResult.success) {
-        return characterResult;
-      }
+    return CharacterServiceStats.getCharacterSummary(characterId, userId);
+  }
 
-      const character = characterResult.data;
-      const summary = character.toSummary();
+  static async calculateSpellcastingStats(
+    characterId: string,
+    userId: string
+  ): Promise<ServiceResult<SpellcastingStats>> {
+    return CharacterServiceStats.calculateSpellcastingStats(characterId, userId);
+  }
 
-      return createSuccessResult(summary);
+  static async calculateCarryingCapacity(
+    characterId: string,
+    userId: string
+  ): Promise<ServiceResult<CarryingCapacity>> {
+    return CharacterServiceStats.calculateCarryingCapacity(characterId, userId);
+  }
 
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.databaseError('get character summary', error)
-      );
-    }
+  static async calculateEquipmentWeight(
+    characterId: string,
+    userId: string
+  ): Promise<ServiceResult<EquipmentWeight>> {
+    return CharacterServiceStats.calculateEquipmentWeight(characterId, userId);
+  }
+
+  static async calculateExperienceInfo(
+    characterId: string,
+    userId: string
+  ): Promise<ServiceResult<ExperienceInfo>> {
+    return CharacterServiceStats.calculateExperienceInfo(characterId, userId);
   }
 
   // ================================
-  // Character Templates and Cloning
+  // Templates and Bulk Operations - Delegate to CharacterServiceTemplates
   // ================================
 
-  /**
-   * Create character template
-   */
   static async createCharacterTemplate(
     characterId: string,
     userId: string,
     templateName: string
   ): Promise<ServiceResult<CharacterPreset>> {
-    try {
-      const characterResult = await this.getCharacterById(characterId, userId);
-      if (!characterResult.success) {
-        return characterResult;
-      }
-
-      const ownershipCheck = await this.checkCharacterOwnership(characterId, userId);
-      if (!ownershipCheck.success) {
-        return ownershipCheck;
-      }
-
-      const character = characterResult.data;
-
-      // Create template from character
-      const template: CharacterPreset = {
-        name: templateName,
-        type: character.type,
-        race: character.race,
-        class: character.classes[0].class,
-        level: character.classes[0].level,
-        abilityScores: character.abilityScores,
-        hitPoints: character.hitPoints.maximum,
-        armorClass: character.armorClass,
-      };
-
-      return createSuccessResult(template);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.templateCreationFailed(error?.message || 'Unknown error')
-      );
-    }
+    return CharacterServiceTemplates.createCharacterTemplate(characterId, userId, templateName);
   }
 
-  /**
-   * Clone character
-   */
   static async cloneCharacter(
     characterId: string,
     userId: string,
     newName: string
   ): Promise<ServiceResult<ICharacter>> {
-    try {
-      const characterResult = await this.getCharacterById(characterId, userId);
-      if (!characterResult.success) {
-        return characterResult;
-      }
-
-      const originalCharacter = characterResult.data;
-
-      // Create clone data
-      const cloneData: CharacterCreation = {
-        name: newName,
-        type: originalCharacter.type,
-        race: originalCharacter.race,
-        customRace: originalCharacter.customRace,
-        size: originalCharacter.size,
-        classes: originalCharacter.classes,
-        abilityScores: originalCharacter.abilityScores,
-        hitPoints: {
-          maximum: originalCharacter.hitPoints.maximum,
-          current: originalCharacter.hitPoints.maximum, // Reset current HP
-          temporary: 0, // Reset temporary HP
-        },
-        armorClass: originalCharacter.armorClass,
-        speed: originalCharacter.speed,
-        proficiencyBonus: originalCharacter.proficiencyBonus,
-        savingThrows: originalCharacter.savingThrows,
-        skills: Object.fromEntries(originalCharacter.skills),
-        equipment: originalCharacter.equipment,
-        spells: originalCharacter.spells,
-        backstory: originalCharacter.backstory,
-        notes: originalCharacter.notes,
-        imageUrl: originalCharacter.imageUrl,
-      };
-
-      return this.createCharacter(userId, cloneData);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.operationFailed('clone character', error?.message)
-      );
-    }
+    return CharacterServiceTemplates.cloneCharacter(characterId, userId, newName);
   }
 
-  /**
-   * Create character from template
-   */
   static async createCharacterFromTemplate(
     templateData: CharacterPreset,
     ownerId: string,
     customizations?: Partial<CharacterCreation>
   ): Promise<ServiceResult<ICharacter>> {
-    try {
-      // Convert template to character creation data
-      const characterData: CharacterCreation = {
-        name: customizations?.name || templateData.name,
-        type: templateData.type,
-        race: templateData.race,
-        size: 'medium',
-        classes: [{
-          class: templateData.class,
-          level: templateData.level,
-          hitDie: this.getHitDieForClass(templateData.class),
-        }],
-        abilityScores: templateData.abilityScores,
-        hitPoints: {
-          maximum: templateData.hitPoints,
-          current: templateData.hitPoints,
-          temporary: 0,
-        },
-        armorClass: templateData.armorClass,
-        speed: 30,
-        proficiencyBonus: Math.ceil(templateData.level / 4) + 1,
-        savingThrows: {
-          strength: false, dexterity: false, constitution: false,
-          intelligence: false, wisdom: false, charisma: false,
-        },
-        skills: {},
-        equipment: [],
-        spells: [],
-        backstory: '',
-        notes: '',
-        ...customizations,
-      };
+    return CharacterServiceTemplates.createCharacterFromTemplate(templateData, ownerId, customizations);
+  }
 
-      return this.createCharacter(ownerId, characterData);
+  static async createMultipleCharacters(
+    ownerId: string,
+    charactersData: CharacterCreation[]
+  ): Promise<ServiceResult<BulkOperationResult<ICharacter>>> {
+    return CharacterServiceTemplates.createMultipleCharacters(ownerId, charactersData);
+  }
 
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.templateCreationFailed(error?.message || 'Invalid template data')
-      );
-    }
+  static async updateMultipleCharacters(
+    userId: string,
+    updates: Array<{ characterId: string; data: CharacterUpdate }>
+  ): Promise<ServiceResult<ICharacter[]>> {
+    return CharacterServiceTemplates.updateMultipleCharacters(userId, updates);
+  }
+
+  static async deleteMultipleCharacters(
+    userId: string,
+    characterIds: string[]
+  ): Promise<ServiceResult<void>> {
+    return CharacterServiceTemplates.deleteMultipleCharacters(userId, characterIds);
   }
 
   // ================================
   // Character Validation and Sanitization
   // ================================
 
-  /**
-   * Validate character data
-   */
   static async validateCharacterData(
     characterData: any
   ): Promise<ServiceResult<CharacterCreation>> {
@@ -788,9 +272,6 @@ export class CharacterService {
   // Character Ownership and Permissions
   // ================================
 
-  /**
-   * Check character ownership
-   */
   static async checkCharacterOwnership(
     characterId: string,
     userId: string
@@ -822,9 +303,6 @@ export class CharacterService {
     }
   }
 
-  /**
-   * Check character access (ownership or public)
-   */
   static async checkCharacterAccess(
     characterId: string,
     userId: string
@@ -851,9 +329,6 @@ export class CharacterService {
     }
   }
 
-  /**
-   * Get character permissions
-   */
   static async getCharacterPermissions(
     characterId: string,
     userId: string
@@ -884,195 +359,8 @@ export class CharacterService {
   }
 
   // ================================
-  // Bulk Operations
-  // ================================
-
-  /**
-   * Create multiple characters
-   */
-  static async createMultipleCharacters(
-    ownerId: string,
-    charactersData: CharacterCreation[]
-  ): Promise<ServiceResult<BulkOperationResult<ICharacter>>> {
-    try {
-      const successful: ICharacter[] = [];
-      const failed: Array<{ data: any; error: string }> = [];
-
-      for (const characterData of charactersData) {
-        const result = await this.createCharacter(ownerId, characterData);
-        if (result.success) {
-          successful.push(result.data);
-        } else {
-          failed.push({
-            data: characterData,
-            error: result.error.message,
-          });
-        }
-      }
-
-      return createSuccessResult({ successful, failed });
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.operationFailed('create multiple characters', error?.message)
-      );
-    }
-  }
-
-  /**
-   * Update multiple characters
-   */
-  static async updateMultipleCharacters(
-    userId: string,
-    updates: Array<{ characterId: string; data: CharacterUpdate }>
-  ): Promise<ServiceResult<ICharacter[]>> {
-    try {
-      const results: ICharacter[] = [];
-
-      for (const update of updates) {
-        const result = await this.updateCharacter(update.characterId, userId, update.data);
-        if (result.success) {
-          results.push(result.data);
-        }
-      }
-
-      return createSuccessResult(results);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.operationFailed('update multiple characters', error?.message)
-      );
-    }
-  }
-
-  /**
-   * Delete multiple characters
-   */
-  static async deleteMultipleCharacters(
-    userId: string,
-    characterIds: string[]
-  ): Promise<ServiceResult<void>> {
-    try {
-      for (const characterId of characterIds) {
-        const result = await this.deleteCharacter(characterId, userId);
-        if (!result.success) {
-          throw new Error(`Failed to delete character ${characterId}: ${result.error.message}`);
-        }
-      }
-
-      return createSuccessResult(void 0);
-
-    } catch (error) {
-      return createErrorResult(
-        CharacterServiceErrors.operationFailed('delete multiple characters', error?.message)
-      );
-    }
-  }
-
-  // ================================
-  // Additional Statistics Methods (Stubs for now)
-  // ================================
-
-  static async calculateSpellcastingStats(
-    _characterId: string,
-    _userId: string
-  ): Promise<ServiceResult<SpellcastingStats>> {
-    // TODO: Implement spellcasting calculations
-    return createSuccessResult({
-      casterLevel: 0,
-      spellSlots: {},
-      spellAttackBonus: 0,
-      spellSaveDC: 8,
-    });
-  }
-
-  static async calculateCarryingCapacity(
-    _characterId: string,
-    _userId: string
-  ): Promise<ServiceResult<CarryingCapacity>> {
-    // TODO: Implement carrying capacity calculations
-    return createSuccessResult({
-      maximum: 240,
-      current: 55,
-      encumbranceLevel: 'none',
-    });
-  }
-
-  static async calculateEquipmentWeight(
-    _characterId: string,
-    _userId: string
-  ): Promise<ServiceResult<EquipmentWeight>> {
-    // TODO: Implement equipment weight calculations
-    return createSuccessResult({
-      total: 55,
-      equipped: 55,
-      carried: 0,
-    });
-  }
-
-  static async calculateExperienceInfo(
-    _characterId: string,
-    _userId: string
-  ): Promise<ServiceResult<ExperienceInfo>> {
-    // TODO: Implement experience calculations
-    return createSuccessResult({
-      currentXP: 0,
-      currentLevel: 1,
-      nextLevelXP: 300,
-      xpToNextLevel: 300,
-      isMaxLevel: false,
-    });
-  }
-
-  // ================================
   // Private Helper Methods
   // ================================
-
-  private static async getCharacterLimitForUser(_userId: string): Promise<number> {
-    // TODO: Implement subscription tier checking
-    return 10; // Default limit for now
-  }
-
-  private static async checkCharacterInUse(_characterId: string): Promise<ServiceResult<void>> {
-    // TODO: Check if character is in active encounters
-    return createSuccessResult(void 0);
-  }
-
-  private static getSkillAbilityModifier(
-    skill: string,
-    abilityModifiers: Record<string, number>
-  ): number {
-    // Simplified skill to ability mapping
-    const skillAbilityMap: Record<string, string> = {
-      athletics: 'strength',
-      acrobatics: 'dexterity',
-      intimidation: 'charisma',
-      // Add more mappings as needed
-    };
-
-    const ability = skillAbilityMap[skill] || 'dexterity';
-    return abilityModifiers[ability] || 0;
-  }
-
-  private static getHitDieForClass(className: string): number {
-    const hitDieMap: Record<string, number> = {
-      barbarian: 12,
-      fighter: 10,
-      paladin: 10,
-      ranger: 10,
-      bard: 8,
-      cleric: 8,
-      druid: 8,
-      monk: 8,
-      rogue: 8,
-      warlock: 8,
-      sorcerer: 6,
-      wizard: 6,
-      artificer: 8,
-    };
-
-    return hitDieMap[className] || 8;
-  }
 
   private static sanitizeText(text: string): string {
     // Basic HTML/XSS sanitization
@@ -1082,3 +370,14 @@ export class CharacterService {
       .trim();
   }
 }
+
+// Re-export types and interfaces for convenience
+export type {
+  PaginatedCharacters,
+  CharacterStats,
+  SpellcastingStats,
+  CarryingCapacity,
+  EquipmentWeight,
+  ExperienceInfo,
+  BulkOperationResult,
+};
